+++ src/screens/detailsScreen.tsx
